split edit component into form component and data component, implemented update method, loops get method
