+++ frontend/src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import './styles/App.css';
-import EditIndexCard from "./component/EditIndexCard";
 import useIndexCard from "./hook/useIndexCard";
 import {ToastContainer} from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css'
 import Header from "./component/Header";
 import IndexCardOverview from "./component/IndexCardOverview";
 import {Route, Routes} from 'react-router-dom';
+import IndexCardForm from "./component/IndexCardForm";
 
 function App() {
 
@@ -20,7 +20,7 @@ function App() {
                 <Route path="/"
                        element={<IndexCardOverview indexCards={indexCards}/>}/>
                 <Route path="/add"
-                       element={<EditIndexCard/>}/>
+                       element={<IndexCardForm/>}/>
             </Routes>
         </div>
     );
