+++ frontend/src/component/IndexCardComponent.tsx
@@ -1,39 +1,42 @@
 import {IndexCard} from "../model/IndexCard";
-import {Box, Card, CardContent, Fab, Typography} from "@mui/material";
-import EditIcon from '@mui/icons-material/Edit';
+import {Button, Card, CardContent} from "@mui/material";
 import '../styles/IndexCardComponent.css';
-import {Route, Routes } from "react-router-dom";
-import EditIndexCard from "./EditIndexCard";
+import IndexCardData from "./IndexCardData";
+import IndexCardForm from "./IndexCardForm";
+import {useState} from "react";
 
 type IndexCardProps = {
     indexCard: IndexCard
 }
 export default function IndexCardComponent({indexCard}: IndexCardProps) {
 
-    return (
-        <div id={"indexCard"}>
-            <Card>
-                <CardContent>
-                    <div id={"topRow"}>
-                        <Box>
-                            <Typography>{indexCard.term1}</Typography>
-                        </Box>
-                        <Box>
-                            <Typography>{indexCard.term2}</Typography>
-                        </Box>
-                        <Box>
-                            <Typography>{indexCard.difficulty}</Typography>
-                        </Box>
-                        <Fab><EditIcon/></Fab>
-                    </div>
-                    <Routes>
-                        <Route path="/edit"
-                               element={
-                            <EditIndexCard indexCard={indexCard}/>
-                        }/>
-                    </Routes>
-                </CardContent>
-            </Card>
-        </div>
-    )
+    const [edit, setEdit] = useState<boolean>(false)
+
+    const switchEdit = () => {
+        setEdit(!edit)
+    }
+
+    const renderedComponent = () => {
+        if (!edit) {
+            return (<div>
+                <IndexCardData indexCard={indexCard}/>
+                <Button onClick={switchEdit}>Edit</Button>
+            </div>)
+        } else {
+            return (
+                <div>
+                    <IndexCardForm indexCard={indexCard}/>
+                    <Button onClick={switchEdit}>Edit</Button>
+                </div>
+            )
+        }
+    }
+
+    return (<Card>
+        <CardContent>
+            {renderedComponent()}
+        </CardContent>
+    </Card>)
+
+
 }
