--- frontend/src/component/IndexCardDetailPage.tsx
@@ -1,11 +1,0 @@
-import IndexCardComponent from "./IndexCardComponent";
-import {IndexCard} from "../model/IndexCard";
-
-type detailsProps = {
-    indexCard: IndexCard
-}
-export default function IndexCardDetailPage({indexCard}: detailsProps) {
-    return (
-        <IndexCardComponent indexCard={indexCard}/>
-    )
-}
